Move polynomial_basis to libhelfem
Also orthpoly.h, which is only used in polynomial_basis.cc
