+++ libhelfem/CMakeLists.txt
@@ -9,6 +9,7 @@ add_library(helfem STATIC
     src/chebyshev.cc
     src/lobatto.cc
     src/polynomial.cc
+    src/polynomial_basis.cc
 )
 target_include_directories(helfem PUBLIC "${CMAKE_CURRENT_BINARY_DIR}/include/" "${ARMADILLO_INCLUDE_DIRS}")
 target_include_directories(helfem PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/src/")
