+++ src/atomic/RadialBasis.h
@@ -17,7 +17,7 @@
 #define ATOMIC_BASIS_RADIALBASIS_H
 
 #include <armadillo>
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 #include "../general/model_potential.h"
 
 namespace helfem {
