+++ src/atomic/TwoDBasis.cpp
@@ -16,7 +16,7 @@
 #include "TwoDBasis.h"
 #include "basis.h"
 #include "quadrature.h"
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 #include "chebyshev.h"
 #include "../general/spherical_harmonics.h"
 #include "../general/gaunt.h"
