+++ src/atomic/basis.h
@@ -17,7 +17,7 @@
 #define ATOMIC_BASIS_H
 
 #include <armadillo>
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 #include "../general/model_potential.h"
 #include "../general/sap.h"
 #include "RadialBasis.h"
