+++ src/atomic/inttest.cpp
@@ -14,7 +14,7 @@
  * of the License, or (at your option) any later version.
  */
 #include "quadrature.h"
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 #include "polynomial.h"
 #include "chebyshev.h"
