+++ src/atomic/main.cpp
@@ -21,7 +21,7 @@
 #include "../general/elements.h"
 #include "../general/timer.h"
 #include "../general/scf_helpers.h"
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 #include "basis.h"
 #include "dftgrid.h"
 #include <cfloat>
