+++ src/diatomic/basis.cpp
@@ -16,7 +16,7 @@
 #include "basis.h"
 #include "quadrature.h"
 #include "polynomial.h"
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 #include "chebyshev.h"
 #include "../general/spherical_harmonics.h"
 #include "../general/gaunt.h"
