+++ src/diatomic/quadrature.h
@@ -18,7 +18,7 @@
 
 #include <armadillo>
 #include "../general/legendretable.h"
-#include "../general/polynomial_basis.h"
+#include "polynomial_basis.h"
 
 namespace helfem {
   namespace diatomic {
